requirements.txt
Flask==2.0.1
Werkzeug==2.0.3
pymongo==3.11.4
python-dotenv==0.19.0
